write_file_from_template() moved to a utils module.
